requirements.txt
aiosignal==1.3.1
asttokens==2.4.1
attrs==23.2.0
boto3==1.34.44
botocore==1.34.44
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
executing==2.0.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
ipykernel==6.29.0
ipython==8.21.0
jedi==0.19.1
jmespath==1.0.1
joblib==1.3.2
JPype1==1.5.0
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.26.3
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
phonenumbers==8.13.29
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
s3transfer==0.10.0
scikit-learn==1.4.0
scipy==1.12.0
setuptools==68.2.2
six==1.16.0
SQLAlchemy==2.0.25
stack-data==0.6.3
tabula-py==2.9.0
threadpoolctl==3.2.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
validate-email==1.3
wcwidth==0.2.13
wheel==0.41.2
yarl==1.9.4
